Rename torrent module to metainfo.
The previously named torrent object was really just a lightweight representation of the metainfo file
