+++ opalescence/ui/__init__.py
@@ -1,6 +1,6 @@
-#!/usr/bin/env python
 # -*- coding: utf-8 -*-
+
 """
-The UI for opalescence is implemented here.
+The UI for opalescence is implemented in this package.
 Currently, only a cli is provided.
 """
